GPU native half precision. Compiles, still doing testing

+++ src/species_advance/standard/energy_p.cc
@@ -245,11 +245,11 @@ energy_p_kernel(const k_interpolator_t& k_interp, const k_particles_soa_t& k_par
         float dy = k_part.dy(n);
         float dz = k_part.dz(n);
         int   i  = k_part.i(n);
-        float v0 = k_part.ux(n) + qdt_2mc*(    ( k_interp(i, interpolator_var::ex)    + dy*k_interp(i, interpolator_var::dexdy)    ) +
+        float v0 = static_cast<float>(k_part.ux(n)) + qdt_2mc*(    ( k_interp(i, interpolator_var::ex)    + dy*k_interp(i, interpolator_var::dexdy)    ) +
                            dz*( k_interp(i, interpolator_var::dexdz) + dy*k_interp(i, interpolator_var::d2exdydz) ) );
-        float v1 = k_part.uy(n) + qdt_2mc*(    ( k_interp(i, interpolator_var::ey)    + dz*k_interp(i, interpolator_var::deydz)    ) +
+        float v1 = static_cast<float>(k_part.uy(n)) + qdt_2mc*(    ( k_interp(i, interpolator_var::ey)    + dz*k_interp(i, interpolator_var::deydz)    ) +
                            dx*( k_interp(i, interpolator_var::deydx) + dz*k_interp(i, interpolator_var::d2eydzdx) ) );
-        float v2 = k_part.uz(n) + qdt_2mc*(    ( k_interp(i, interpolator_var::ez)    + dx*k_interp(i, interpolator_var::dezdx)    ) +
+        float v2 = static_cast<float>(k_part.uz(n)) + qdt_2mc*(    ( k_interp(i, interpolator_var::ez)    + dx*k_interp(i, interpolator_var::dezdx)    ) +
                                 dy*( k_interp(i, interpolator_var::dezdy) + dx*k_interp(i, interpolator_var::d2ezdxdy) ) );
         v0 = v0*v0 + v1*v1 + v2*v2;
         v0 = (msp * k_part.w(n)) * (v0 / (1 + sqrtf(1 + v0)));
